requirements.txt
flask==2.3.3
langchain==0.0.340
groq>=0.5.0
python-docx==0.8.11
pdfplumber==0.10.3
python-dotenv==1.0.0
